ModelClassesToTsClassesConverter extracted from ModelClassesToTsAngular2JsonApiConverter
